Created basic GUI, will come back later to make it more presentable. Also, started on the "check" functions
Renamed my existing .py files to fit the MVC format.  Finally figured out the GUI and made it fairly presentable. Did all the calculations for RT60 and plotted the frequencies. If I'm being honest, I basically redid my project to have a controller.py because it was working with just two .py files and added classes(which are not my favorite). I also did forget to commit so that is why this one is so long and its so late. I'm just figuring out these last few bugs and it should be ready to turn in(even if it doesn't do everything, its gotta count for something right?)